ユーザーID列の型を`Long`に変更し、IDP作成ロジックや関連メソッドを実装。`VerifiedToken`のプロパティ名を変更して一貫性を確保。

+++ ktse/src/main/kotlin/net/kigawa/keruta/ktse/persist/db/table/UserIdpTable.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object UserIdpTable: Table("user_idp") {
-    val userId = integer("user_id").references(UserTable.id)
+    val userId = long("user_id").references(UserTable.id)
     val issuer = varchar("issuer", 50)
     val subject = varchar("subject", 50)
     val audience = varchar("audience", 50)
